Moved theme.js file and refactored imports to reflect new structure
