+++ client/src/components/landing-components/Hero.jsx
@@ -4,7 +4,7 @@ import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Grid';
 import Box from '@mui/material/Box';
 import { Button } from '@mui/material';
-import { tokens } from "../../theme";
+import { tokens } from "../../context/theme";
 import { useTheme } from "@emotion/react";
 import useColorTheme from "../../hooks/FormStyles";
 import { useNavigate } from 'react-router-dom';
